library/express-configuration/pyproject.toml
[tool.poetry]
name = "express-configuration"
version = "1.0.3"
description = ""
authors = ["yjsong <yjsong@interezen.co.kr>"]
readme = "README.md"
packages = [{ include = "express_configuration" }]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
watchdog = "^4.0.1"

[[tool.poetry.source]]
name = "pypi-group"
url = "http://192.168.124.238:8001/repository/pypi-group/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



